# Add semi/anti join support in sort-merge join