This PR adds the ability to execute semi and anti joins in the sort merge join. Main motivation was an observation that came up during @SvenLehmann 's master thesis. In some cases, we have joins where the sides cannot be switched (e.g., certain outer joins or semi/anti) and in these cases the hash join performs poorly when the probe side is much smaller that the build side.
This limitation does not exist for the sort-merge join.

**On top of that, this PR includes some additional changes:** 
- a saner determination of cluster counts which improve performance significantly for small joins
- significantly improve performance by not merging pos lists in the end of the operator which saves runtime and further ensure that each output chunk is guaranteed to be sorted (exploitable by the current seminar groups), which was not the case before
- much more constness, again surprising performance gains
- some simplifications, e.g., removing the dictionary specialisations for the materialization phase (bucket sort is not faster than a final std::sort)
- make skipped tests more explicit by using `GTEST_SKIP()` over returns; we skipped join tests for most join implementations where columns types differed; no these tests are skipped and thus show that we lack support for certain joins

**Todo:**
- `src/lib/operators/join_sort_merge/radix_cluster_sort.hpp:158` does some ugly passing of vectors which I don't like but I am not able to improve that code. Any help is welcome here.

**Benchmarks:**

Here is a simple join benchmark for a `lineitem.l_orderkey` join on SF=1. The hash join is still the fastest join, but when the sides cannot be switched, the sort-merge join can be 2x faster.
This PR does not address this problem from the optimizer side.
```
-----------------------------------------------------------------------------------------------------------------
Benchmark                                                                          Time           CPU Iterations
-----------------------------------------------------------------------------------------------------------------
TPCHDataMicroBenchmarkFixture/BM_HashSemiProbeRelationSmaller_mean         516987105 ns  509959200 ns          1
TPCHDataMicroBenchmarkFixture/BM_HashSemiProbeRelationLarger_mean          234356279 ns  230752325 ns          4
TPCHDataMicroBenchmarkFixture/BM_SortMergeSemiProbeRelationSmaller_mean    254436540 ns  253390633 ns          3
TPCHDataMicroBenchmarkFixture/BM_SortMergeSemiProbeRelationLarger_mean     292223744 ns  291456767 ns          3
TPCHDataMicroBenchmarkFixture/BM_HashAntiProbeRelationSmaller_mean         411283087 ns  409554000 ns          1
TPCHDataMicroBenchmarkFixture/BM_HashAntiProbeRelationLarger_mean          167975552 ns  167341800 ns          4
TPCHDataMicroBenchmarkFixture/BM_SortMergeAntiProbeRelationSmaller_mean    265472312 ns  264762950 ns          2
TPCHDataMicroBenchmarkFixture/BM_SortMergeAntiProbeRelationLarger_mean     332750246 ns  329975550 ns          2
```

This is TPC-H SF=1 on nemea. On the left is the current master that always uses the sort-merge join. On the right is this PR.
```
+-----------+--------------------+------+--------------------+------+--------+---------------------------------+
| Benchmark | prev. iter/s       | runs | new iter/s         | runs | change | p-value (significant if <0.001) |
+-----------+--------------------+------+--------------------+------+--------+---------------------------------+
| TPC-H 1   | 0.582531273365021  | 35   | 0.599179029464722  | 36   | +3%    |                          0.0002 |
| TPC-H 2   | 2.61675119400024   | 158  | 2.71726107597351   | 164  | +4%    |                          0.0000 |
| TPC-H 3   | 3.4964451789856    | 210  | 3.78179144859314   | 227  | +8%    |                          0.0000 |
| TPC-H 4   | 2.92605233192444   | 176  | 3.19961619377136   | 193  | +9%    |                          0.0000 |
| TPC-H 5   | 1.42285799980164   | 86   | 1.53936815261841   | 93   | +8%    |                          0.0000 |
| TPC-H 6   | 25.2097625732422   | 1513 | 25.3011417388916   | 1519 | +0%    |                          0.0063 |
| TPC-H 7   | 0.736438393592834  | 45   | 0.701482534408569  | 43   | -5%    |                          0.0000 |
| TPC-H 8   | 1.83221352100372   | 110  | 1.88769721984863   | 114  | +3%    |                          0.0000 |
| TPC-H 9   | 0.858280956745148  | 52   | 0.91465151309967   | 55   | +7%    |                          0.0000 |
| TPC-H 10  | 2.32912540435791   | 140  | 2.37053990364075   | 143  | +2%    |                          0.0000 |
| TPC-H 11  | 11.1193714141846   | 668  | 12.1195287704468   | 728  | +9%    |                          0.0000 |
| TPC-H 12  | 5.2208104133606    | 314  | 5.30656576156616   | 319  | +2%    |                          0.0000 |
| TPC-H 13  | 1.82378625869751   | 110  | 1.96606266498566   | 118  | +8%    |                          0.0000 |
| TPC-H 14  | 15.5952310562134   | 936  | 15.7704248428345   | 947  | +1%    |                          0.0000 |
| TPC-H 15  | 10.3300695419312   | 620  | 12.0110139846802   | 721  | +16%   |                          0.0000 |
| TPC-H 16  | 6.6278920173645    | 398  | 6.72749662399292   | 404  | +2%    |                          0.0000 |
| TPC-H 17  | 0.146589562296867  | 9    | 0.146263644099236  | 9    | -0%    |        (not enough runs) 0.9572 |
| TPC-H 18  | 1.37537503242493   | 83   | 1.37091338634491   | 83   | -0%    |                          0.5590 |
| TPC-H 19  | 5.32353925704956   | 320  | 5.38209962844849   | 323  | +1%    |                          0.0000 |
| TPC-H 20  | 0.0759870260953903 | 5    | 0.0779949873685837 | 5    | +3%    |        (not enough runs) 0.1984 |
| TPC-H 21  | 0.122050642967224  | 8    | 0.122789047658443  | 8    | +1%    |        (not enough runs) 0.8324 |
| TPC-H 22  | 9.37542819976807   | 563  | 11.0998754501343   | 666  | +18%   |                          0.0000 |
| average   |                    |      |                    |      | +4%    |                                 |
+-----------+--------------------+------+--------------------+------+--------+---------------------------------+
```

The changes mainly impact very small joins which do not a significant impact on TPC-H. Might be different for TPC-C and alike.

Fixes #1499.