@Bouncner 2019-02-19
Fixes #1499.